Rename server pkg to runner
